requirements.txt
asgiref==3.5.0
Django==4.0.3
django-filter==23.2
djangorestframework==3.14.0
importlib-metadata==6.8.0
Markdown==3.4.4
pytz==2023.3.post1
sqlparse==0.4.2
zipp==3.16.2
